+++ packages/terra-functional-testing/src/config/utils/getConfigurationOptions.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const getCapabilities = require('./getCapabilities');
 const getIpAddress = require('./getIpAddress');
+const getDefaultThemeName = require('./getDefaultThemeName');
 
 const defaultWebpackPath = path.resolve(process.cwd(), 'webpack.config.js');
 
@@ -10,6 +11,7 @@ const getConfigurationOptions = (options) => {
     assetServerPort,
     browsers,
     disableSeleniumService,
+    ignoreScreenshotMismatch,
     externalHost,
     externalPort,
     formFactor,
@@ -21,22 +23,25 @@ const getConfigurationOptions = (options) => {
     suite,
     theme,
     updateScreenshots,
+    useSeleniumStandaloneService,
   } = options;
 
   return {
     baseUrl: `http://${externalHost || getIpAddress()}:${externalPort || 8080}`,
     capabilities: getCapabilities(browsers, !!gridUrl),
-    hostname: gridUrl || 'localhost',
+    hostname: gridUrl || (useSeleniumStandaloneService ? 'standalone-chrome' : 'localhost'),
     port: gridUrl ? 80 : 4444,
     launcherOptions: {
-      disableSeleniumService: disableSeleniumService || !!gridUrl,
+      disableSeleniumService: disableSeleniumService || useSeleniumStandaloneService || !!gridUrl,
+      ignoreScreenshotMismatch,
       formFactor,
       gridUrl,
       keepAliveSeleniumDockerService,
       locale,
       port: assetServerPort,
       site,
-      theme,
+      ...(theme ? { theme } : { theme: getDefaultThemeName() }),
+      overrideTheme: theme,
       updateScreenshots,
       ...fs.existsSync(defaultWebpackPath) && { webpackConfig: defaultWebpackPath },
     },