+++ packages/terra-functional-testing/src/terra-cli/wdio/test-runner.js
@@ -3,6 +3,7 @@ const path = require('path');
 const Launcher = require('@wdio/cli').default;
 const { Logger } = require('@cerner/terra-cli');
 const getConfigurationOptions = require('../../config/utils/getConfigurationOptions');
+const { cleanScreenshots, downloadScreenshots } = require('../../commands/utils');
 
 const logger = new Logger({ prefix: '[terra-functional-testing:wdio]' });
 
@@ -45,6 +46,7 @@ class TestRunner {
    * @param {array} options.suite - Overrides specs and runs only the defined suites.
    * @param {string} options.theme - A theme for the test run.
    * @param {boolean} options.updateScreenshots - Updates all reference screenshots with the latest screenshots.
+   * @param {boolean} options.useSeleniumStandaloneService - A flag to use the selenium standalone service instead of the selenium docker service.
    * @returns {Promise} A promise that resolves with the test run exit code.
    */
   static async run(options) {
@@ -85,6 +87,7 @@ class TestRunner {
    * @param {array} options.suite - Overrides specs and runs only the defined suites.
    * @param {string} options.themes - A list of themes for the test run.
    * @param {boolean} options.updateScreenshots - Updates all reference screenshots with the latest screenshots.
+   * @param {boolean} options.useSeleniumStandaloneService - A flag to use the selenium standalone service instead of the selenium docker service.
    */
   static async start(options) {
     const {
@@ -93,6 +96,11 @@ class TestRunner {
       themes,
       ...cliOptions
     } = options;
+
+    // Clean only the non reference screenshots.
+    cleanScreenshots();
+
+    await TestRunner.configureScreenshots(options);
 
     /**
      * The following code loops through each permutation of theme, locale, and form factor.
@@ -119,6 +127,31 @@ class TestRunner {
           formFactorIndex += 1;
         } while (formFactorIndex < formFactors.length);
       }
+    }
+  }
+
+  static async configureScreenshots(options) {
+    const {
+      config,
+      screenshotUrl,
+    } = options;
+
+    let url = screenshotUrl;
+
+    if (!url) {
+      const configPath = TestRunner.configPath(config);
+      // eslint-disable-next-line global-require, import/no-dynamic-require
+      const wdioConfig = require(configPath);
+      const { screenshots } = wdioConfig.config;
+
+      if (screenshots) {
+        url = screenshots.url;
+      }
+    }
+
+    if (url) {
+      logger.info(`Starting to download screenshots from ${url}.`);
+      await downloadScreenshots(url);
     }
   }
 }