+++ packages/terra-functional-testing/tests/jest/commands/utils/describeTests.test.js
@@ -1,5 +1,4 @@
-const describeViewports = require('../../../../src/commands/utils/describeViewports');
-const describeTests = require('../../../../src/commands/utils/describeTests');
+const { describeViewports, describeTests } = require('../../../../src/commands/utils');
 
 jest.mock('../../../../src/commands/utils/describeViewports');