+++ packages/terra-functional-testing/tests/jest/commands/utils/getViewports.test.js
@@ -1,4 +1,4 @@
-const getViewports = require('../../../../src/commands/utils/getViewports');
+const { getViewports } = require('../../../../src/commands/utils');
 const { TERRA_VIEWPORTS } = require('../../../../src/constants');
 
 const {