+++ packages/terra-functional-testing/tests/jest/commands/utils/setViewport.test.js
@@ -1,4 +1,4 @@
-const setViewport = require('../../../../src/commands/utils/setViewport');
+const { setViewport } = require('../../../../src/commands/utils');
 const { TERRA_VIEWPORTS } = require('../../../../src/constants');
 
 const mockSetWindowSize = jest.fn();